Move InhibitionRule-related models to separate file
